LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/redirect/multiple-redirect-https-downgrade-upgrade.sub-expected.txt

Downgraded then upgraded font


Harness Error (TIMEOUT), message = null

PASS Https downgrade-upgrade iframe
PASS Https downgrade-upgrade top level navigation
TIMEOUT Https downgrade-upgrade embed Test timed out
NOTRUN Https downgrade-upgrade fetch() api
NOTRUN Https downgrade-upgrade object
PASS Https downgrade-upgrade preload
NOTRUN Https downgrade-upgrade stylesheet
NOTRUN Https downgrade-upgrade track
NOTRUN Https downgrade-upgrade image => No headers
PASS Https downgrade-upgrade script => No headers
FAIL Https downgrade-upgrade script => No headers: sec-fetch-dest assert_equals: expected "script" but got ""
FAIL Https downgrade-upgrade script => No headers: sec-fetch-mode assert_equals: expected "no-cors" but got ""
FAIL Https downgrade-upgrade script => No headers: sec-fetch-site assert_equals: expected "cross-site" but got ""
PASS Https downgrade-upgrade script => No headers: sec-fetch-user
FAIL Https downgrade-upgrade iframe: sec-fetch-dest assert_equals: expected "iframe" but got ""
FAIL Https downgrade-upgrade iframe: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL Https downgrade-upgrade iframe: sec-fetch-site assert_equals: expected "cross-site" but got ""
PASS Https downgrade-upgrade iframe: sec-fetch-user
NOTRUN Https downgrade-upgrade font => No headers
FAIL Https downgrade-upgrade top level navigation: sec-fetch-dest assert_equals: expected "document" but got ""
FAIL Https downgrade-upgrade top level navigation: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL Https downgrade-upgrade top level navigation: sec-fetch-site assert_equals: expected "cross-site" but got ""
PASS Https downgrade-upgrade top level navigation: sec-fetch-user
